backend/requirements.txt
lask==2.1.2
flask-cors==3.0.10
psycopg2-binary==2.9.3